+++ src/main/java/com/evilduck/filecatcher/service/FileService.java
@@ -24,7 +24,7 @@ public abstract class FileService {
 
     public abstract String save(InputStream inputStream, String fileName, String contentType) throws IOException;
 
-    public abstract String saveOrAppend(InputStream inputStream, String fileName, int startByte, int totalFileBytes, String contentType) throws IOException;
+    public abstract String saveOrAppend(InputStream inputStream, String fileName, long startByte, long totalFileBytes, String contentType) throws IOException;
 
 
     public abstract void process(List<String> jobIds);
